+++ afs/parsers.py
@@ -1,5 +1,4 @@
 import os
-import ujson
 import ast
 import json
 from pathlib import Path
@@ -106,12 +105,12 @@ def manifest_parser(notebook_path, pypi_endpoint, output_dir=None, manifest_yaml
     requirements = '--index-url {0}\n--trusted-host {1}\n'.format(pypi_endpoint, pypi_host) + requirements
 
     if 'requirements' in manifest:
-        requirements_append = [req for req in manifest['requirements'] if req not in req_list]
+        requirements_append = list(set([req for req in manifest['requirements'] if req not in req_list]))
 
         if afs_sdk_version:
-            for afs_req in [req for req in requirements_append if req.startswith('afs')]:
-                requirements_append.remove(afs_req)
-            requirements_append.append('afs=={}'.format(afs_sdk_version))
+            if 'afs' in requirements_append:
+                requirements_append.remove('afs')
+                requirements_append.append('afs=={}'.format(afs_sdk_version))
 
         requirements = requirements + '\n'.join(requirements_append)
 
@@ -181,7 +180,7 @@ def config_to_dict(source, startswith='node_config'):
         config = source
 
     try:
-        config = ujson.loads(config)
+        config = json.loads(config)
     except ValueError:
         try:
             config = ast.literal_eval(config)